requirements.txt
alembic==1.18.4
annotated-types==0.7.0
dnspython==2.8.0
Mako==1.3.12
MarkupSafe==3.0.3
psycopg==3.3.4
psycopg-binary==3.3.4
pydantic_core==2.46.4
python-dotenv==1.2.2
SQLAlchemy==2.0.49
typing-inspection==0.4.2
typing_extensions==4.15.0
